Report response times
