+++ src/main/scala/com.github.loadtest4j.drivers.gatling/GatlingResult.scala
@@ -3,9 +3,10 @@ package com.github.loadtest4j.drivers.gatling
 import java.time.Duration
 import java.util.Optional
 
-import com.github.loadtest4j.loadtest4j.DriverResult
+import com.github.loadtest4j.loadtest4j.ResponseTime
+import com.github.loadtest4j.loadtest4j.driver.DriverResult
 
-class GatlingResult(ok: Long, ko: Long, actualDuration: Duration, reportUrl: String) extends DriverResult {
+class GatlingResult(ok: Long, ko: Long, actualDuration: Duration, responseTime: ResponseTime, reportUrl: String) extends DriverResult {
   override def getKo: Long = ko
 
   override def getOk: Long = ok
@@ -13,4 +14,6 @@ class GatlingResult(ok: Long, ko: Long, actualDuration: Duration, reportUrl: Str
   override def getReportUrl: Optional[String] = Optional.of(reportUrl)
 
   override def getActualDuration: Duration = actualDuration
+
+  override def getResponseTime: ResponseTime = responseTime
 }
